+++ NEWS.md
@@ -1,3 +1,5 @@
+# idiolect (development version)
+
 # idiolect 1.0.1
 
 * Fixed issues after CRAN review.