+++ NEWS.md
@@ -1,3 +1,5 @@
+# idiolect 1.1.1
+
 # idiolect 1.1.0
 
 * minor bug fixes